Create and use more specific decoding error exceptions
I'm deriving DecodingError (now an abstract base class) into new,
specific exception classes so that the user can get raw exception
properties instead of a single reason string (for example, when the
iterator is moving beyond the current packet's content size, what was
the requested size and the current packet content's size).

The reason string is still available in DecodingError::reason(). It is
cooked from the exception properties in each specific exception's
constructor.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>