*.adoc: convert to Asciidoctor format
As of this date, GitHub uses Asciidoctor to render a `.adoc` file. There
are a few minor differences between the legacy Asciidoc and new
Asciidoctor formats which need to be addressed. In yactfr's files, they
are:

* $$C++$$ becomes pass:[C++] (double dollar is deprecated in
  Asciidoctor).
* No double underscore within a literal block; Asciidoctor supports
  normal text formatting within single ticks.
* Use `+something+` for non-formatted monospace text.
* Asciidoctor does not support ``something'' for double quotes: use
  "`something`".

I added a comment at the beginning of each file to indicate that you
should render them with Asciidoctor.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>