Clock type: require that offset in cycles be less than frequency
This makes sense: the part in seconds of the offset is used to
compensate the offset in cycles because the frequency can be very high.
For example, given a frequency of 1000 Hz, the following are equivalent:

    Offset (s):           45     48
    Offset (cycles):    3423    423

This adds a guarantee and can simplify some timestamp computations for
tools using yactfr.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>