TSDL parser: support having `stream` block after `event` block needing it
LTTng-UST seems to produce this.

The solution is to add pseudo event record types as orphans instead of
trying to add them to an already created pseudo data stream type. When
creating the trace type, we use the pseudo orphan event record types to
create event record types for the appropriate, matching data stream
type.

Tests are added to make sure this continues to work.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>