Iterator: skip padding before setting a `ScopeBeginningElement`
With a layout such as:

    event {
        context := struct {
            integer { align = 32; size = 32; } a;
            integer { align = 8; size = 8; } b;
        };
        fields := struct {
            integer { align = 32; size = 32; } c;
            integer { align = 32; size = 32; } d;
        };
    };

Pretend the offset of field `a` offset is 0, then the field offsets
(bits) are:

    a     0
    b    32
    c    64
    d    96

Without aligning for the payload structure before setting a
`ScopeBeginningElement`, the iterator's offset() method returns 40 for
this element. For trace inspection purposes, it is more useful that
offset() always returns the beginning of the effective data, excluding
any leading padding. In this example, the offset() method must return 64
for the second `ScopeBeginningElement` element.

Graphically:

    aaaabb**ccccdddd
    //////########## <- payload scope beginning to end (before)
    //////  ######## <- payload scope beginning to end (now)
      ^- context scope (unchanged)

Exists this somewhat weird scenario:

    event {
        context := struct {
            enum : integer { align = 32; size = 32; } {
                OPT,
            } t;
        };
        fields := struct {
            variant <event.context.tag> {
                integer { align = 64; size = 32; } OPT;
            } v;
        };
    };

Because the variant `v` does not affect the payload structure's
alignment, there's still padding at the beginning of said structure:

        ============ <- payload structure beginning to end
    tttt****vvvvvvvv
    ////############ <- payload scope beginning to end

But I believe this is okay: since the padding bits are located within
the payload structure, this padding is actually within the scope.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>