yactfr/CMakeLists.txt: do not use `Boost::boost` imported target
It looks like, for some versions of CMake (tested on CMake 3.5.1), the
`Boost::boost` imported target is not created by the `FindBoost` module
when no components are specified in the find_package() call.

Instead of "linking" to this imported target just to get the Boost
include directories, include `${Boost_INCLUDE_DIRS}` within
target_include_directories().

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>