Metadata text test: remove `uuid` field when there's no `uuid` property
I consider this test to be invalid here because the packet header
contains a `uuid` field, but there's no UUID property to compare to
during an eventual decoding process.

This test fails (but it should pass) before this patch because of the
recent trace type's optional UUID property fix. It is the only test with
this pattern.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>