Fix: TSDL parsing: parse square brackets in reverse order
I have no words.

For some reason,

    string allo[2][len][4]

was parsed as a static array of four dynamic arrays of `len` static
arrays of two strings. But Babeltrace (1 and 2... 😒) and the C language
understand a static array of two dynamic arrays of `len` static arrays
of four strings.

Here's a concrete example:

    event {
      name = ev;
      fields := struct {
        integer { size = 8; } a[2][3];
      };
    };

Babeltrace 1 prints:

    0 ev: { a = [ [0] = [ [0] = 17, [1] = 34, [2] = 51 ],
                  [1] = [ [0] = 68, [1] = 85, [2] = 102 ] ] }

Even the tests verified this wrong behaviour in yactfr.

This is just another proof that I'm still the only user of this very
general-purpose C++14 library.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>