Make internal::JsonParser::_tryParseNumber() more efficient
This patch makes internal::JsonParser::_tryParseNumber() try to parse
constant integers first, and then check if the parsed integer could in
fact be a constant real, in which case it backtracks and tries to parse
a constant real.

This is more efficient because internal::StrScanner::tryScanConstReal()
always calls std::regex_search() to confirm the constant JSON real
number form.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>