tests/tests-metadata-text: validate the error messages of expect-to-fail tests
All those tests used to simply check that yactfr::fromMetadataText()
throws (which `metadata-text-tester` converted to returning the exit
status 2).

Now, the files contain both the metadata stream text and the expected
error messages (which `metadata-text-tester` now prints to the standard
output, exiting with status 0).

I used `metadata-text-tester` itself to append those expected messages
to the metadata stream files, and then manually validated that the
expected error messages match the test intentions. In doing so, I found
a few tests which failed for the wrong reason and fixed them.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>