Fix: decode valid 10-byte variable-length integers
This patch changes the VM so that it can decode valid 10-byte
variable-length integers.

Without this patch, the VM throws
`OversizedVariableLengthIntegerDecodingError` when a variable-length
integer has more than nine bytes. However, the following examples are
valid:

Unsigned:
    ff ff ff ff ff ff ff ff ff 01:
        The value 18,446,744,073,709,551,615.

Signed:
    ff ff ff ff ff ff ff ff ff 00:
        The value 9,223,372,036,853,775,807.

    80 80 80 80 80 80 80 80 80 7f:
        The value -9,223,372,036,854,775,808.

Because 9 × 7 is 63, the 10th byte, if any, must contain a single
effective MSB to fit a 64-bit variable. When the variable-length integer
is signed, said bit is the sign bit of a 64-bit signed integer.

My strategy might not be optimal here as it relies on conditional
statements:

* One in _appendVlIntByte() to validate some conditions (that's where we
  throw `OversizedVariableLengthIntegerDecodingError`) and to make sure
  `_pos.curVlIntLenBits` becomes 64 (our maximum) instead of 70
  (10 × 7).

  Although something like

      81 80 80 80 80 80 80 80 80 80 80 80 80 80 80 80 80 80 80 80 00

  should be valid (the value is 1), yactfr doesn't support it as of this
  version. In other words, above nine bytes, the variable-length integer
  must be as compact as possible.

* One (new) unfortunate in _signExtendVlSIntVal() because

      static_cast<std::uint64_t>(1) << _pos.curVlIntLenBits

  with `_pos.curVlIntLenBits` being 64 is UB.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>