Fix: make assert() in `IntegerRange` ctor. conditional
Consider the following CTF 2 metadata stream input:

    {
      "type": "optional",
      "selector-field-location": ["event-record-payload", "sel"],
      "selector-field-ranges": [[-23, 17]],
      "field-class": {
        "type": "null-terminated-string"
      }
    }

This `[-23, 17]` integer range leads to pseudoDtFromOptWithIntSelFc()
creating an integer range set of `unsigned long long`, casting `-23`
to `unsigned long long`, leading to this failed assertion in the
constructor of `IntegerRange`:

    assert(lower <= upper);

This is a documented, internal hack:

    /*
     * This is a hack: such an integer range set may in fact contain
     * signed integer ranges, but we'll only know in
     * DtFromPseudoRootDtConverter::_dtFromPseudoVarWithIntRangesType(),
     * at which point signed integer range sets will be created, casting
     * the lower and upper values to `long long`.
     *
     * Hackish, but safe.
     */

The right solution, if I had more time, would be to use pseudo integer
range and integer range set classes for the pseudo type phase, and then
convert such instances to yactfr ones once we know the true underlying
integer type.

But this patch is a hack itself, adding a `ValidatePreconditionsV`
template parameter to the `IntegerRange` and `IntegerRangeSet` class
templates so that the assert() call doesn't exist during the pseudo type
phase.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>