Add bit order property to FL bit array types (CTF2‑SPECRC‑9.0)
This patch adds the bit order property to fixed-length bit array types.

The bit order controls the filling direction of the theoretical output
bit array. In CTF 2, the two values are "first to last" and "last to
first".

The bit order is closely related to the byte order, in that there's a
natural filling direction for a given byte order, and that's the default
in CTF 2 as well as in the yactfr API (the fixed-length bit array types
accept an optional bit order, setting it according to the byte order if
missing):

Big endian:
    Last to first

Little endian:
    First to last

The natural filling direction is what yactfr (and all CTF decoders) did
and are doing to this point. The bit order property makes the encoding
and decoding even more flexible. For example, given the two following
bytes (binary):

    11010010 10101001

Given an aligned decoding position, the resulting bit arrays are:

Big endian, last to first (natural):
    1 0 0 1 0 1 0 1 0 1 0 0 1 0 1 1

Big endian, first to last:
    1 1 0 1 0 0 1 0 1 0 1 0 1 0 0 1

Little endian, first to last (natural):
    0 1 0 0 1 0 1 1 1 0 0 1 0 1 0 1

Little endian, last to first:
    1 0 1 0 1 0 0 1 1 1 0 1 0 0 1 0

Note that in all the results above, the first element is consider to be
the LSB of a corresponding integer. For example, in the big endian and
last to first example, the LSB is 1, and then bit 1 is 0, bit 2 is 0,
and so on.

In the internals of yactfr, the unnatural filling direction is called
"reversed". Therefore, the new `*_REV` procedure instructions instruct
the VM to reverse the decoded fixed-length bit array. Whether or not to
reverse is a template parameter of various helpers in `vm.hpp`;
therefore, this patch doesn't affect the decoding performance for
existing cases.

The bit reversal function is in `yactfr/internal/fl-int-rev.hpp`. I used
this blog post [1] to perform it.

I also considered an alternative approach to implement this feature: not
performing any bit reversal and documenting as such. Then the user may
perform a reversal herself when getting a fixed-length bit array element
thanks to the byte and bit order properties of the type.

However, I believe the purpose of yactfr is to provide true values, and
fixed-length integer and floating point number elements would be wrong
with an unnatural filling direction.

Moreover, an unnatural filling direction is a very unusual and rare
case, so I believe the additional bit reversal cost is acceptable.
Anyway, whether or not to perform bit reversals could be configurable
per fixed-length bit array type or at the trace type level in the
future.

Tests are minimal at this point, but it seems to work.

[1]: https://matthewarcus.wordpress.com/2012/11/18/reversing-a-64-bit-word/

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>