tests: use moultipart (from Babeltrace 2) and Normand
This patch replaces the custom text-to-binary format of the yactfr tests
with Normand [1].

It also replaces the custom multi-file format (with the `----`
delimiter) with the moultipart format of the Babeltrace 2 [2] testing
framework.

Less custom things here means less maintenance and free features from
other projects.

I converted all the `tests/tests-iter-data` stream files to Normand
manually, adding more structure and comments as I did.

Also changed in this patch is that a test which fails when some output
isn't expected shows a diff to easily spot the unexpected lines.

[1]: https://github.com/efficios/normand
[2]: https://babeltrace.org/

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>